app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pg8000==1.31.2
sqlalchemy==2.0.31
cloud-sql-python-connector[pg8000]==1.11.0
google-cloud-alloydb-connector[pg8000]==1.1.0
google-cloud-aiplatform==1.58.0
vertexai==1.58.0
google-auth==2.32.0
